:tada: feat: static code analysis on trace
